+++ examples/example.py
@@ -4,44 +4,74 @@ import dimstack as ds
 ds.display.mode("rich")
 
 k = 0.25
-target_process_sigma = 6
+target_process_sigma = 3
 stdev = 0.036 / target_process_sigma
-m1 = ds.dim.Statistical(
-    nom=208,
-    tol=ds.tol.SymmetricBilateral(0.036),
+m1 = ds.dim.Reviewed(
+    dim=ds.dim.Basic(
+        nom=208,
+        tol=ds.tol.SymmetricBilateral(0.036),
+        name="a",
+        desc="Shaft",
+    ),
     distribution=ds.dist.Normal(208 + k * target_process_sigma * stdev, stdev),
     target_process_sigma=target_process_sigma,
-    name="a",
-    desc="Shaft",
 )
-m2 = ds.dim.Statistical(
-    nom=-1.75,
-    tol=ds.tol.UnequalBilateral(0, 0.06),
+m2 = ds.dim.Reviewed(
+    dim=ds.dim.Basic(
+        nom=-1.75,
+        tol=ds.tol.UnequalBilateral(0, 0.06),
+        name="b",
+        desc="Retainer ring",
+    ),
     target_process_sigma=3,
-    name="b",
-    desc="Retainer ring",
 )
-m3 = ds.dim.Statistical(nom=-23, tol=ds.tol.UnequalBilateral(0, 0.12), target_process_sigma=3, name="c", desc="Bearing")
-m4 = ds.dim.Statistical(
-    nom=20,
-    tol=ds.tol.SymmetricBilateral(0.026),
+m3 = ds.dim.Reviewed(
+    dim=ds.dim.Basic(
+        nom=-23,
+        tol=ds.tol.UnequalBilateral(0, 0.12),
+        name="c",
+        desc="Bearing",
+    ),
     target_process_sigma=3,
-    name="d",
-    desc="Bearing Sleeve",
 )
-m5 = ds.dim.Statistical(nom=-200, tol=ds.tol.SymmetricBilateral(0.145), target_process_sigma=3, name="e", desc="Case")
+m4 = ds.dim.Reviewed(
+    dim=ds.dim.Basic(
+        nom=20,
+        tol=ds.tol.SymmetricBilateral(0.026),
+        name="d",
+        desc="Bearing Sleeve",
+    ),
+    target_process_sigma=3,
+)
+m5 = ds.dim.Reviewed(
+    dim=ds.dim.Basic(
+        nom=-200,
+        tol=ds.tol.SymmetricBilateral(0.145),
+        name="e",
+        desc="Case",
+    ),
+    target_process_sigma=3,
+)
 m6 = ds.dim.Basic(
     nom=20,
     tol=ds.tol.SymmetricBilateral(0.026),
-    # target_process_sigma=3,
     name="f",
     desc="Bearing Sleeve",
 )
-m7 = ds.dim.Statistical(nom=-23, tol=ds.tol.UnequalBilateral(0, 0.12), target_process_sigma=3, name="g", desc="Bearing")
-items = [m1, m2, m3, m4, m5, m6, m7]
+m7 = ds.dim.Reviewed(
+    dim=ds.dim.Basic(
+        nom=-23,
+        tol=ds.tol.UnequalBilateral(0, 0.12),
+        name="g",
+        desc="Bearing",
+    ),
+    target_process_sigma=3,
+)
+items = [m1, m2, m3, m4, m5, m7]
 
-stack = ds.Stack(name="stacks on stacks", dims=items)
+stack = ds.ReviewedStack(name="stacks on stacks", dims=items)
 
+stack.to_basic_stack().show()
 stack.show()
 ds.calc.Closed(stack).show()
 ds.calc.WC(stack).show()
@@ -50,7 +80,7 @@ ds.calc.MRSS(stack).show()
 designed_for = ds.calc.SixSigma(stack, at=4.5)
 designed_for.show()
 
-spec = ds.Spec("stack spec", "", distribution=designed_for.distribution, LL=0.05, UL=0.8)
+spec = ds.Requirement("stack spec", "", distribution=designed_for.distribution, LL=0.05, UL=0.8)
 spec.show()
 
-# ds.plot.StackPlot().add(stack).add(stack.RSS).show()
+ds.plot.StackPlot().add(stack).add(ds.calc.RSS(stack)).show()