+++ src/dimstack/__init__.py
@@ -1,7 +1,7 @@
 # from . import eval
 from . import calc, dim, display, dist, plot, stats, tolerance, utils
 from . import tolerance as tol
-from .dim import Basic, Spec, Stack, Statistical
+from .dim import Basic, BasicStack, Requirement, Reviewed, ReviewedStack
 from .dist import Normal, Uniform  # , Triangular, LogNormal, Weibull, Exponential, Gamma, Beta, Gumbel, Frechet
 
 __all__ = ["dim", "stats", "display", "tolerance", "utils", "dist", "plot", "calc"]