+++ src/dimstack/calc.py
@@ -1,15 +1,18 @@
-from typing import List
-
-from .dim import Basic, Stack, Statistical
+from .dim import Basic, BasicStack, Reviewed, ReviewedStack
 from .stats import rss
 from .tolerance import Bilateral
 
 
-def Closed(self: Stack) -> Basic:
-    nominal = sum([dim.dir * dim.nominal * dim.a for dim in self.dims])
+def Closed(self: BasicStack | ReviewedStack) -> Basic:
+    if isinstance(self, BasicStack):
+        dims = self.dims
+    elif isinstance(self, ReviewedStack):
+        dims = [rdim.dim for rdim in self.dims]
+
+    nominal = sum([dim.dir * dim.nominal * dim.a for dim in dims])
     tolerance = Bilateral(
-        sum(filter(None, [dim.abs_upper_tol for dim in self.dims])),
-        sum(filter(None, [dim.abs_lower_tol for dim in self.dims])),
+        sum(filter(None, [dim.abs_upper_tol for dim in dims])),
+        sum(filter(None, [dim.abs_lower_tol for dim in dims])),
     )
     return Basic(
         nominal,
@@ -19,13 +22,19 @@ def Closed(self: Stack) -> Basic:
     )
 
 
-def WC(self: Stack) -> Basic:
+def WC(self: BasicStack | ReviewedStack) -> Basic:
     """
     This is a simple WC calculation. This results in a Bilateral dimension with a tolerance that is the sum of the component tolerances.
     It states that in any combination of tolerances, you can be sure the result will be within the this resulting tolerance.
     """
-    mean = sum([dim.dir * dim.median * dim.a for dim in self.dims])
-    t_wc = sum([abs((dim.tolerance.T / 2) * dim.a) for dim in self.dims])
+
+    if isinstance(self, BasicStack):
+        dims = self.dims
+    elif isinstance(self, ReviewedStack):
+        dims = [rdim.dim for rdim in self.dims]
+
+    mean = sum([dim.dir * dim.median * dim.a for dim in dims])
+    t_wc = sum([abs((dim.tolerance.T / 2) * dim.a) for dim in dims])
     tolerance = Bilateral(t_wc)
     return Basic(
         nom=mean,
@@ -35,7 +44,7 @@ def WC(self: Stack) -> Basic:
     )
 
 
-def RSS(self: Stack) -> Statistical:
+def RSS(self: BasicStack | ReviewedStack) -> Basic:
     """
     This is a simple RSS calculation. This is uses the RSS calculation method in the Dimensioning and Tolerancing Handbook, McGraw Hill.
     It is really only useful for a Bilateral stack of same process-stdev dims. The RSS result has the same uncertainty as the measurements.
@@ -54,25 +63,33 @@ def RSS(self: Stack) -> Statistical:
         - Dimensioning and Tolerancing Handbook, McGraw Hill
         - http://files.engineering.com/getfile.aspx?folder=69759f43-e81a-4801-9090-a0c95402bfc0&file=RSS_explanation.GIF
     """
-    dims: List[Statistical] = [Statistical.from_basic_dimension(dim) for dim in self.dims]
+    if isinstance(self, BasicStack):
+        dims = self.dims
+    elif isinstance(self, ReviewedStack):
+        dims = [rdim.dim for rdim in self.dims]
+
     d_g = sum([dim.dir * dim.median * dim.a for dim in dims])
     t_rss = rss([dim.dir * (dim.tolerance.T / 2) * dim.a for dim in dims])
     tolerance = Bilateral(t_rss)
-    return Statistical(
+    return Basic(
         nom=d_g,
         tol=tolerance,
         name=f"{self.name} - RSS Analysis",
         desc="(assuming inputs with Normal Dist. & uniform SD)",
-    ).assume_normal_dist()
+    )
 
 
-def MRSS(self: Stack) -> Statistical:
+def MRSS(self: BasicStack | ReviewedStack) -> Basic:
     """Basically RSS with a coefficient modifier to make the tolerance tighter.
 
     Returns:
         Statistical: _description_
     """
-    dims: List[Statistical] = [Statistical.from_basic_dimension(dim) for dim in self.dims]
+    if isinstance(self, BasicStack):
+        dims = self.dims
+    elif isinstance(self, ReviewedStack):
+        dims = [rdim.dim for rdim in self.dims]
+
     d_g = sum([dim.dir * dim.median * dim.a for dim in dims])
     t_wc = sum([abs(dim.dir * (dim.tolerance.T / 2) * dim.a) for dim in dims])
     t_rss = rss([dim.dir * dim.a * (dim.tolerance.T / 2) for dim in dims])
@@ -80,28 +97,26 @@ def MRSS(self: Stack) -> Statistical:
     C_f = (0.5 * (t_wc - t_rss)) / (t_rss * (n**0.5 - 1)) + 1
     t_mrss = C_f * t_rss
     tolerance = Bilateral(t_mrss)
-    stdev = t_wc / 6
-    sigma = t_mrss / stdev
-    return Statistical(
+    return Basic(
         nom=d_g,
         tol=tolerance,
         name=f"{self.name} - MRSS Analysis",
         desc="(assuming inputs with Normal Dist. & uniform SD)",
-        target_process_sigma=sigma,
-    ).assume_normal_dist()
-
-
-def SixSigma(self: Stack, at: float = 3) -> Statistical:
-    dims: List[Statistical] = [Statistical.from_basic_dimension(dim).assume_normal_dist() for dim in self.dims]
-    mean = sum([dim.dir * dim.median for dim in dims])
-    stdev = rss([dim.stdev_eff for dim in dims])
-    tolerance = Bilateral(stdev * at)
-    dim = Statistical(
-        nom=mean,
-        tol=tolerance,
-        target_process_sigma=at,
-        name=f"{self.name} - '6 Sigma' Analysis",
-        desc="(assuming inputs with Normal Dist.)",
     )
-    dim.assume_normal_dist()
+
+
+def SixSigma(self: ReviewedStack, at: float = 3) -> Reviewed:
+    # mean = sum([rdim.mean_eff for rdim in self.dims])
+    mean = sum([rdim.dim.dir * rdim.dim.median for rdim in self.dims])
+    stdev = rss([dim.stdev_eff for dim in self.dims])
+    tolerance = Bilateral(stdev * at)
+    dim = Reviewed(
+        Basic(
+            nom=mean,
+            tol=tolerance,
+            name=f"{self.name} - '6 Sigma' Analysis",
+            desc="(assuming inputs with Normal Dist.)",
+        ),
+        target_process_sigma=at,
+    )
     return dim