+++ tests/test_distribution_types.py
@@ -12,20 +12,22 @@ class PositiveDist(unittest.TestCase):
         measurements = np.array([1.1, 0.9])
         measurements_dist = dimstack.dist.Normal.fit(measurements)
         tol = dimstack.tol.SymmetricBilateral(0.3)
-        dim = dimstack.dim.Statistical(
-            nom=1,
-            tol=tol,
+        rdim = dimstack.dim.Reviewed(
+            dim=dimstack.dim.Basic(
+                nom=1,
+                tol=tol,
+                name="1",
+            ),
             distribution=measurements_dist,
-            name="1",
         )
 
-        self.assertEqual(dim.nominal, 1)
-        self.assertEqual(dim.abs_lower, 0.7)
-        self.assertEqual(dim.abs_upper, 1.3)
+        self.assertEqual(rdim.dim.nominal, 1)
+        self.assertEqual(rdim.dim.abs_lower, 0.7)
+        self.assertEqual(rdim.dim.abs_upper, 1.3)
 
         self.assertEqual(measurements_dist.mean, 1)
 
-        self.assertAlmostEqual(float(dim.yield_probability), 0.9973, 4)
+        self.assertAlmostEqual(float(rdim.yield_probability), 0.9973, 4)
 
 
 class NegativeDist(unittest.TestCase):
@@ -33,20 +35,22 @@ class NegativeDist(unittest.TestCase):
         measurements = np.array([-1.1, -0.9])
         measurements_dist = dimstack.dist.Normal.fit(measurements)
         tol = dimstack.tol.SymmetricBilateral(0.3)
-        dim = dimstack.dim.Statistical(
-            nom=-1,
-            tol=tol,
+        rdim = dimstack.dim.Reviewed(
+            dim=dimstack.dim.Basic(
+                nom=-1,
+                tol=tol,
+                name="1",
+            ),
             distribution=measurements_dist,
-            name="1",
         )
 
-        self.assertEqual(dim.nominal * dim.dir, -1)
-        self.assertEqual(dim.abs_lower, -1.3)
-        self.assertEqual(dim.abs_upper, -0.7)
+        self.assertEqual(rdim.dim.nominal * rdim.dim.dir, -1)
+        self.assertEqual(rdim.dim.abs_lower, -1.3)
+        self.assertEqual(rdim.dim.abs_upper, -0.7)
 
         self.assertEqual(measurements_dist.mean, -1)
 
-        self.assertAlmostEqual(float(dim.yield_probability), 0.9973, 4)
+        self.assertAlmostEqual(float(rdim.yield_probability), 0.9973, 4)
 
 
 if __name__ == "__main__":